# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KKN Candles API",
        "description": "API pour la boutique de bougies KKN Candles",
        "contact": {
            "name": "Support KKN Candles",
            "email": "support@kkncandles.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/admin/products": {
            "get": {
                "tags": [
                    "Admin - Products"
                ],
                "summary": "Liste des produits (Admin)",
                "description": "Récupérer tous les produits avec filtrage et pagination",
                "operationId": "adminGetProducts",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filtrer par catégorie",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filtrer par statut actif",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Rechercher par nom ou slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre de produits par page (max 100)",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des produits",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Product"
                                                    }
                                                },
                                                "meta": {
                                                    "$ref": "#/components/schemas/PaginationMeta"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Products"
                ],
                "summary": "Créer un produit",
                "description": "Créer un nouveau produit",
                "operationId": "adminCreateProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "name",
                                    "slug",
                                    "price",
                                    "stock_quantity"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Bougie Parfumée Vanille"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "bougie-parfumee-vanille"
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 1000,
                                        "nullable": true
                                    },
                                    "long_description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "price": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 5000
                                    },
                                    "stock_quantity": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 100
                                    },
                                    "in_stock": {
                                        "type": "boolean",
                                        "default": true
                                    },
                                    "is_featured": {
                                        "type": "boolean",
                                        "default": false
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "default": true
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "url"
                                        },
                                        "nullable": true
                                    },
                                    "specifications": {
                                        "type": "object",
                                        "nullable": true
                                    },
                                    "sort_order": {
                                        "type": "integer",
                                        "default": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produit créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Produit créé avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Modèle produit",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "category_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "name": {
                        "type": "string",
                        "example": "Bougie Parfumée Vanille"
                    },
                    "slug": {
                        "type": "string",
                        "example": "bougie-parfumee-vanille"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "long_description": {
                        "type": "string",
                        "nullable": true
                    },
                    "price": {
                        "type": "integer",
                        "example": 5000
                    },
                    "formatted_price": {
                        "type": "string",
                        "example": "5 000 FCFA"
                    },
                    "stock_quantity": {
                        "type": "integer",
                        "example": 100
                    },
                    "in_stock": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": false
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "url"
                        }
                    },
                    "main_image": {
                        "type": "string",
                        "format": "url",
                        "nullable": true
                    },
                    "specifications": {
                        "type": "object",
                        "nullable": true
                    },
                    "sort_order": {
                        "type": "integer",
                        "example": 0
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Category": {
                "title": "Category",
                "description": "Modèle catégorie",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string",
                        "example": "Bougies Parfumées"
                    },
                    "slug": {
                        "type": "string",
                        "example": "bougies-parfumees"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "format": "url",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "sort_order": {
                        "type": "integer",
                        "example": 0
                    },
                    "products_count": {
                        "type": "integer",
                        "example": 15
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "title": "PaginationMeta",
                "description": "Métadonnées de pagination",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 12
                    },
                    "total": {
                        "type": "integer",
                        "example": 120
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification JWT Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}